Refactor codebase
Refactored the existing code to improve structure and maintainability. Consider reviewing specific components for further optimization.
[skip gpt_engineer]
