+++ src/components/ui/sidebar/types.ts
@@ -1,0 +1,37 @@
+import { VariantProps } from "class-variance-authority"
+import { sidebarMenuButtonVariants } from "./variants"
+
+export type SidebarState = "expanded" | "collapsed"
+
+export type SidebarContext = {
+  state: SidebarState
+  open: boolean
+  setOpen: (open: boolean) => void
+  openMobile: boolean
+  setOpenMobile: (open: boolean) => void
+  isMobile: boolean
+  toggleSidebar: () => void
+}
+
+export interface SidebarProviderProps {
+  defaultOpen?: boolean
+  open?: boolean
+  onOpenChange?: (open: boolean) => void
+  className?: string
+  style?: React.CSSProperties
+  children: React.ReactNode
+}
+
+export interface SidebarProps extends React.ComponentProps<"div"> {
+  side?: "left" | "right"
+  variant?: "sidebar" | "floating" | "inset"
+  collapsible?: "offcanvas" | "icon" | "none"
+}
+
+export interface SidebarMenuButtonProps
+  extends React.ComponentProps<"button">,
+    VariantProps<typeof sidebarMenuButtonVariants> {
+  asChild?: boolean
+  isActive?: boolean
+  tooltip?: string | React.ComponentProps<"div">
+}
